requirements.txt
discord==1.7.3
discord.py==1.7.3
peewee==3.14.9
python-dotenv
